requirements.txt
groq>=0.4.0
langgraph>=0.2.0
streamlit>=1.30.0
python-docx>=1.0.0
mlcroissant>=1.0.0
python-dotenv>=1.0.0
